requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

numpy==2.1.3
faiss-cpu==1.9.0
sentence-transformers==3.3.1

pydantic==2.10.3

httpx==0.28.1
beautifulsoup4==4.12.3

pandas==2.2.3
openpyxl==3.1.5




fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.11.2
httpx==0.28.1

pandas==2.2.3
openpyxl==3.1.5

numpy<2
scikit-learn==1.5.2
faiss-cpu==1.9.0.post1

rank-bm25==0.2.2

sentence-transformers==3.3.1

streamlit==1.41.1
requests==2.32.3
